Fix bad <em> tags
Tonc seemingly predates <em> being commonly used for italics, and instead used it for big glaring 'temp' or 'broken' style indicators.
This is a problem for conversion because most markup languages will produce <em> when they see italics.
So I've replaced the existing <em> with a .emphasis class and adjusted the styling so that <em> now looks the same as <i>