Port "Chapter 19: Text Systems" to markdown
This commit migrates `text.htm` to `text.md`, and converts a few things
to play nicely with the mdbook styling.

Most notably, the figure showing bitpacking of fonts required a fair bit
of custom CSS. This was done with an in-line style tag to take advantage
of more easily selecting the relevant dom elements.