Port "Chapter 19: Text Systems" to markdown (#60)
* Port "Chapter 19: Text Systems" to markdown

This commit migrates `text.htm` to `text.md`, and converts a few things
to play nicely with the mdbook styling.

Most notably, the figure showing bitpacking of fonts required a fair bit
of custom CSS. This was done with an in-line style tag to take advantage
of more easily selecting the relevant dom elements.

* deprecation notice box

* more boxes converted

* Update styles for the bit unpacking table.

The bit unpacking table was not clear in many of the mdbook themes. On
coranac.com, it was styled more akin the the bit packing figure, so I
made it leverage those styles as well.

---------

Co-authored-by: Antonio Vivace <avivace4@gmail.com>