Port Mode 7 to Markdown
Lots of equations in this one. I'm like 85% confident I ported them over
correctly, but getting some of the alignment that Cearn was doing with
tables was quite tricky enough to the point where I didn't fully bother.
If you know of a way of introducing significant whitespace in MathML,
we can follow up and correct some alignment quirks.