Port Mode7 Part 2 to Markdown. (#70)
This commit migrates `mode7ex.htm` to `mode7ex.md`. There were a large
number of equations in this page. For one or two of them, I slightly
adjusted how the equation was written (without breaking the math, of
course), for better clarity when represented with MathML vs styling a
`<table>` element.