Port the Changelog to markdown
The only thing of note with this conversion is tonc's usage of custom
bullet points on list elements. In order to preserve keeping the
majority of the page in markdown, we instead discussed in the Discord
server to keep the lists written in markdown, and add the images in the
content.